[CVW-049] Feat, UserConfig레포지토리 대역명 수정
